backend/pyproject.toml
[project]
name = "kindalike-backend"
version = "0.1.0"
description = "Backend API for KindaLike restaurant recommender"
authors = [
    {name = "Hemanjali Reddy"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "langchain (>=1.2.3,<2.0.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "langchain-core (>=1.2.7,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "loguru (>=0.7.0,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
